Home Page Options Cleanup
